+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Removed stale _logger object
 - Fixed Smoothing (#1863)
 - Fixed Crash when XCB,X11 was configured and display manager changed to Wayland
+- Fixed Target-directory not correctly build when exporting effects
 
 **JSON-API**
 - Refactored JSON-API to ensure consistent authorization behaviour across sessions and single requests with token authorization.